dla: fix suspicious entry detection on each scan
Firstly we move the task to stuck list, thus task always
belongs to some list, so next check on suspicious is always
true.

If task is always suspicious that means it was successfully
analysed on task scan and this task has correct syscall
and we can do stack unwinding.

But if task was marked as need_relax it won't be put to
any stuck list, but still it can be found in the whole
stuck chain, hence ptrace can fail (syscall was never
correctly set).

To avoid this we have to be sure, that if not suspicious task
(which does not belong to any stuck list) is found, chain
is broken, we break out of the loop and the whole scan must be
repeated.

Signed-off-by: Roman Pen <r.peniaev@gmail.com>