dla: remove need_relax hence it breaks the chain
in case of need_relax we skip task checking and we do not
put it into stuck list, thus chain is being broken.

all we need is to consider impact of ptrace syscall, which
changes switch counters and requires some relax after
ptrace has been finished.

solution is simple: update all the counters and wait for task
to return into stuck state.

Signed-off-by: Roman Pen <r.peniaev@gmail.com>